Bump `base` and `time`
