+++ testutil-core/build.gradle.kts
@@ -28,9 +28,11 @@ import io.spine.gradle.internal.Deps
 
 val spineBaseVersion: String by extra
 
+group = "io.spine.tools"
+
 dependencies {
     api(project(":client"))
-    api("io.spine:spine-testlib:$spineBaseVersion")
+    api("io.spine.tools:spine-testlib:$spineBaseVersion")
     Deps.test.apply {
         junit.api.forEach { api(it) }
         truth.libs.forEach { api(it) }
