+++ version.gradle.kts
@@ -40,13 +40,13 @@
 /**
  * Version of this library.
  */
-val coreJava = "2.0.0-SNAPSHOT.4"
+val coreJava = "2.0.0-SNAPSHOT.5"
 
 /**
  * Versions of the Spine libraries that `core-java` depends on.
  */
-val base = "2.0.0-SNAPSHOT.4"
-val time = "2.0.0-SNAPSHOT.4"
+val base = "2.0.0-SNAPSHOT.5"
+val time = "2.0.0-SNAPSHOT.5"
 
 project.extra.apply {
     this["versionToPublish"] = coreJava
